Remote locking (#21)
* Added initial measurements

* Upgrade to latest WAVM

* Small tidy-up

* Remove learning rate decay

* Added synchronous config

* Set up for distributed test

* Upped prewarm count

* Switched off full async

* Switch off full async

* Added fibonacci function

Changed partial segment syncing to be in bulk

* Added redis locking

* Added remote locking around partial state updates

* Removed synchronous weight push from step function

* Hacked in no queuing for sgd_step

* Switched to full async config

* Moving to distributed testing

* Fixed double/ int issue in "finished" function

* Fixed some holes in synchronisation

* Actually increment retry count

* Actually fixed sgd_finished

+++ src/data/reuters_parse.cpp
@@ -49,7 +49,7 @@ void parseReutersData(const path &downloadDir, const path &outputDir) {
                 // Split up index:value part
                 const std::vector<std::string> valueTokens = util::tokeniseString(thisToken, ':');
                 int feature = std::stoi(valueTokens[0]);
-                double weight = std::stof(valueTokens[1]);
+                double weight = std::stod(valueTokens[1]);
 
                 // Add to matrix
                 triplets.emplace_back(Eigen::Triplet<double>(feature, exampleCount, weight));
